Improve API
