+++ app/helpers/universal_renderer/ssr_helpers.rb
@@ -7,5 +7,9 @@ module UniversalRenderer
     def ssr_body
       "<!-- SSR_BODY -->".html_safe
     end
+
+    def sanitize_ssr(html)
+      sanitize(html, scrubber: SsrScrubber.new)
+    end
   end
 end
